Structured code by separating UserCard.js
